feat(f3): forest-cli f3 certs list
